Move utils.py into src, move args into args.py, and use args and utils in eval.py
