+++ scripts/eval.py
@@ -5,23 +5,18 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from dataclasses import dataclass
 import tyro
-import imageio
-from IPython.display import HTML
 import torch
 import numpy as np
-import typing
 
 from libero.libero import get_libero_path
-from stable_baselines3.common.vec_env import DummyVecEnv
-from stable_baselines3.common.monitor import Monitor
 from stable_baselines3 import PPO, SAC
 
-from src.envs_gymapi import LowDimensionalObsGymEnv, AgentViewGymEnv, AgentViewGymGoalEnv, LowDimensionalObsGymGoalEnv
-
 import src.patch
+from src.utils import setup_envs, obs_to_video
+from src.args import AlgArgs, EnvArgs
 
 @dataclass
-class Args:
+class Args(EnvArgs, AlgArgs):
     # User specific arguments
     seed: int = None
     """random seed for reproducibility"""
@@ -37,46 +32,6 @@ class Args:
     """if passed in, the custom path will be used for bddl file as opposed to libero default files"""
     bddl_file_name: str = "libero_90/KITCHEN_SCENE6_close_the_microwave.bddl"
     """file name of the BDDL file"""
-    setup_demo_path: str = None
-    """If passed in, runs the actions in the given demonstration before every episode to setup the scene"""
-    shaping_reward: bool = True
-    """if toggled, shaping reward will be off for all goal states"""
-    sparse_reward: float = 10.0
-    """total sparse reward for success"""
-    reward_geoms: str = None
-    """if geoms are passed, those specific geoms will be rewarded, for single object predicates only [format example: ketchup_1_g1,ketchup_1_g2]"""
-    dense_reward_multiplier: float = 1.0
-    """multiplies the last goal state's shaping reward"""
-    steps_per_episode: int = 250
-    """number of steps in episode. If truncate is True, the episode will terminate after this value"""
-
-    # Algorithm specific arguments
-    alg: str = "ppo"
-    """algorithm to use for training"""
-    her: bool = False
-    """if toggled, SAC will use HER otherwise it would not"""
-    num_envs: int = 1
-    """number of LIBERO environments"""
-    visual_observation: bool = False
-    """if toggled, the environment will return visual observation otherwise it would not"""
-
-def obs_to_video(images, filename):
-    """
-    converts a list of images to video and writes the file
-    """
-    video_writer = imageio.get_writer(filename, fps=60)
-    for image in images:
-        video_writer.append_data(image[::-1])
-    video_writer.close()
-    HTML("""
-        <video width="640" height="480" controls>
-            <source src="output.mp4" type="video/mp4">
-        </video>
-        <script>
-            var video = document.getElementsByTagName('video')[0];
-            video.playbackRate = 2.0; // Increase the playback speed to 2x
-            </script>    
-    """)
 
 
 if __name__ == "__main__":
@@ -85,43 +40,13 @@ if __name__ == "__main__":
 
     if args.custom_bddl_path is not None:
         task_name = os.path.basename(args.custom_bddl_path)
-        env_args = {
-            "bddl_file_name": args.custom_bddl_path,
-            "camera_heights": 512,
-            "camera_widths": 512,
-        }
+        bddl_file = args.custom_bddl_path
     else:
-        bddl_file_base = get_libero_path("bddl_files")
         task_name = args.bddl_file_name
-        env_args = {
-            "bddl_file_name": os.path.join(bddl_file_base, task_name),
-            "camera_heights": 512,
-            "camera_widths": 512,
-        }
-
-    # set up reward geoms
-    reward_geoms = args.reward_geoms.split(",") if args.reward_geoms is not None else None
+        bddl_file = os.path.join(get_libero_path("bddl_files"), task_name)
 
     print("Setting up environment")
-    vec_env_class = DummyVecEnv
-    if args.visual_observation:
-        if args.her:
-            envs = vec_env_class(
-                [lambda: Monitor(AgentViewGymGoalEnv(**env_args)) for _ in range(args.num_envs)]
-            )
-        else:
-            envs = vec_env_class(
-                [lambda: Monitor(AgentViewGymEnv(**env_args)) for _ in range(args.num_envs)]
-            )
-    else:
-        if args.her:
-            envs = vec_env_class(
-                [lambda: Monitor(LowDimensionalObsGymGoalEnv(**env_args)) for _ in range(args.num_envs)]
-            )
-        else:
-            envs = vec_env_class(
-                [lambda: Monitor(LowDimensionalObsGymEnv(args.shaping_reward, args.sparse_reward, reward_geoms, args.dense_reward_multiplier, args.steps_per_episode, setup_demo=args.setup_demo_path, **env_args)) for _ in range(args.num_envs)]
-            )
+    envs = setup_envs(bddl_file, args)
 
     # Seeding everything
     if args.seed is not None:
