+++ scripts/train.py
@@ -15,7 +15,8 @@ from libero.libero import get_libero_path
 from stable_baselines3.common.callbacks import CheckpointCallback
 from src.callbacks import VideoWriter
 
-from utils import WandbArgs, AlgArgs, EnvArgs, setup_envs, setup_model
+from src.utils import setup_envs, setup_model
+from src.args import WandbArgs, AlgArgs, EnvArgs
 
 @dataclass
 class Args(WandbArgs, AlgArgs, EnvArgs):
